Refactor publication test data
Subdirectories within test can have their own conftest.py module to
additional fixture data and test functions. These overlay on conftest
modules in directories above.

This commit moves our publications test dataset into conftest so it can
be reused by other tests. Similarly the fixture data that was in
test.publish.data has been moved into the test.publish.conftest and the
test function parameters have been adjusted to load the appropriate fixtures.
