+++ test/conftest.py
@@ -7,6 +7,7 @@ from sqlalchemy_utils import create_database, database_exists, drop_database
 from rialto_airflow.database import create_schema, engine_setup
 from rialto_airflow.schema.harvest import (
     Author,
+    Funder,
     Publication,
     pub_author_association,
 )
@@ -158,3 +159,274 @@ def test_reports_session(test_reports_engine):
         yield sessionmaker(engine_setup("rialto_reports_test", echo=True))
     finally:
         close_all_sessions()
+
+
+@pytest.fixture
+def dim_json():
+    return {
+        "type": "article",
+        "doi": "10.000/000001",
+        "journal": {"title": "Delicious Limes Journal of Science"},
+        "issue": "12",
+        "pages": "1-10",
+        "volume": "1",
+        "mesh_terms": ["Delicions", "Limes"],
+        "pmid": "123",
+        "linkout": "https://example_dim.com",
+    }
+
+
+@pytest.fixture
+def dim_json_no_title():
+    return {"other_data": {"bogus": "no journal title here"}}
+
+
+@pytest.fixture
+def openalex_json():
+    return {
+        "id": "https://openalex.org/W123456789",
+        "type": "preprint",
+        "biblio": {"issue": "11", "first_page": "1", "last_page": "9", "volume": "2"},
+        "primary_location": {
+            "source": {
+                "type": "journal",
+                "display_name": "Ok Limes Journal of Science",
+            }
+        },
+        "locations": [
+            {"pdf_url": "https://example_openalex_pdf.com"},
+        ],
+        "mesh": [
+            {"descriptor_name": "Ok"},
+            {"descriptor_name": "Lemons"},
+        ],
+        "ids": {
+            "doi": "10.000/000001",
+            "pmid": "1234",
+        },
+    }
+
+
+@pytest.fixture
+def openalex_no_title_json():
+    return {
+        "locations": [
+            {
+                "bogus": {
+                    "nope": "not here",
+                }
+            },
+            {"source": {"type": "geography", "display_name": "New York"}},
+        ],
+    }
+
+
+@pytest.fixture
+def wos_json():
+    return {
+        "UID": "WOS:000123456789",
+        "fullrecord_metadata": {
+            "normalized_doctypes": {"doctype": ["Article", "Abstract"]}
+        },
+        "static_data": {
+            "summary": {
+                "pub_info": {
+                    "pubyear": "2023",
+                    "issue": "10",
+                    "vol": "3",
+                    "page": {"begin": "1", "end": "8"},
+                },
+                "titles": {
+                    "count": 2,
+                    "title": [
+                        {"type": "source", "content": "Meh Limes Journal of Science"},
+                        {"type": "bogus", "content": "Bogus"},
+                    ],
+                },
+            }
+        },
+        "dynamic_data": {
+            "cluster_related": {
+                "identifiers": {
+                    "identifier": [
+                        {
+                            "type": "pmid",
+                            "value": "1234567",
+                        }
+                    ],
+                },
+            }
+        },
+    }
+
+
+@pytest.fixture
+def wos_json_no_page_info():
+    return {
+        "static_data": {
+            "summary": {
+                "pub_info": {
+                    "page": {"page_count": "7"},
+                },
+            }
+        }
+    }
+
+
+@pytest.fixture
+def sulpub_json():
+    return {
+        "journal": {
+            "name": "Bad Limes Journal of Science",
+            "issue": "9",
+            "pages": "1-7",
+            "volume": "4",
+        },
+        "pmid": "123456",
+    }
+
+
+@pytest.fixture
+def pubmed_json():
+    return {
+        "MedlineCitation": {
+            "Article": {
+                "ArticleTitle": "Example Title",
+            },
+        },
+        "PubmedData": {
+            "ArticleIdList": {
+                "ArticleId": [
+                    {"@IdType": "pubmed", "#text": "36857419"},
+                    {"@IdType": "doi", "#text": "10.1182/bloodadvances.2022008893"},
+                ]
+            },
+        },
+    }
+
+
+@pytest.fixture
+def dataset(test_session, dim_json, openalex_json, wos_json, sulpub_json, pubmed_json):
+    """
+    This fixture will create two publications, four authors, and two funders.
+    It is designed to test the various types of reports we want to output, where
+    sometimes we want all the publications, and others we want the unique
+    publications by school and department.
+
+    The first publication is authored by all 4 authors, and funded by both
+    funders. The second publication is authored by the first author, and funded by
+    the first funder.
+
+    The first two authors are from the Department of Social Sciences in the
+    School of Humanities and Sciences. The last two authors are both from the
+    School of Engineering, but one is in the Department of Mechanical Engineering,
+    and the other is in Electric Enginering.
+    """
+
+    with test_session.begin() as session:
+        pub = Publication(
+            doi="10.000/000001",
+            title="My Life",
+            apc=123,
+            open_access="gold",
+            pub_year=2023,
+            dim_json=dim_json,
+            openalex_json=openalex_json,
+            wos_json=wos_json,
+            sulpub_json=sulpub_json,
+            pubmed_json=pubmed_json,
+        )
+
+        pub2 = Publication(
+            doi="10.000/000002",
+            title="My Life Part 2",
+            apc=500,
+            open_access="green",
+            pub_year=2024,
+            dim_json=dim_json,
+            openalex_json=openalex_json,
+            wos_json=wos_json,
+            sulpub_json=sulpub_json,
+            pubmed_json=pubmed_json,
+        )
+
+        author1 = Author(
+            first_name="Jane",
+            last_name="Stanford",
+            sunet="janes",
+            cap_profile_id="1234",
+            orcid="0298098343",
+            primary_school="School of Humanities and Sciences",
+            primary_dept="Social Sciences",
+            primary_role="faculty",
+            schools=[
+                "Vice Provost for Undergraduate Education",
+                "School of Humanities and Sciences",
+            ],
+            departments=["Inter-Departmental Programs", "Social Sciences"],
+            academic_council=True,
+        )
+
+        author2 = Author(
+            first_name="Leland",
+            last_name="Stanford",
+            sunet="lelands",
+            cap_profile_id="12345",
+            orcid="02980983434",
+            primary_school="School of Humanities and Sciences",
+            primary_dept="Social Sciences",
+            primary_role="staff",
+            schools=[
+                "School of Humanities and Sciences",
+            ],
+            departments=["Social Sciences"],
+            academic_council=False,
+        )
+
+        author3 = Author(
+            first_name="Frederick",
+            last_name="Olmstead",
+            sunet="folms",
+            cap_profile_id="123456",
+            orcid="02980983422",
+            primary_school="School of Engineering",
+            primary_dept="Mechanical Engineering",
+            primary_role="faculty",
+            schools=["School of Engineering"],
+            departments=["Mechanical Engineering"],
+            academic_council=False,
+        )
+
+        author4 = Author(
+            first_name="Frederick",
+            last_name="Terman",
+            sunet="fterm",
+            cap_profile_id="1234567",
+            orcid="029809834222",
+            primary_school="School of Engineering",
+            primary_dept="Electrical Engineering",
+            primary_role="faculty",
+            schools=["School of Engineering"],
+            departments=["Electrical Engineering"],
+            academic_council=True,
+        )
+
+        funder1 = Funder(
+            name="National Institutes of Health", grid_id="12345", federal=True
+        )
+        funder2 = Funder(
+            name="Andrew Mellon Foundation", grid_id="123456", federal=False
+        )
+
+        pub.authors.append(author1)
+        pub.authors.append(author2)
+        pub.authors.append(author3)
+        pub.authors.append(author4)
+        pub.funders.append(funder1)
+        pub.funders.append(funder2)
+
+        pub2.authors.append(author1)
+        pub2.funders.append(funder1)
+
+        session.add(pub)
+        session.add(pub2)
